+++ client/components/PlaceOrder.js
@@ -1,4 +1,6 @@
 import React from 'react'
+import {Link, withRouter} from 'react-router-dom'
+
 import Fab from '@material-ui/core/Fab'
 import ShoppingBasketOutlined from '@material-ui/icons/ShoppingBasketOutlined'
 
@@ -6,10 +8,12 @@ class PlaceOrder extends React.Component {
   render() {
     const {cart} = this.props
     return (
-      <Fab variant="extended" color="primary" disabled={cart.length === 0}>
-        <ShoppingBasketOutlined />
-        Place Order
-      </Fab>
+      <Link to="/checkout">
+        <Fab variant="extended" color="primary" disabled={cart.length === 0}>
+          <ShoppingBasketOutlined />
+          Checkout
+        </Fab>
+      </Link>
     )
   }
 }
